add 'useStyled()' hook to implement 'styled()' behavior for shadow DOM
This encapsulates all the functionality for the `styled` HOC into a single
hook, making the bundle size impact smaller.
